feat(angular-services): move vehicle service to services folder
